move BnfFirstNextAnalyzer, ParserGeneratorUtil, and RuleGraphHelper to generator module, keep some methods in bnf-language
