+++ bnf-language/src/org/intellij/grammar/psi/impl/GrammarPsiImplUtil.java
@@ -21,7 +21,6 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 
-import static org.intellij.grammar.generator.ParserGeneratorUtil.findRuleImplMethods;
 import static org.intellij.grammar.psi.BnfAttributes.getAttribute;
 import static org.intellij.grammar.psi.BnfAttributes.getRootAttribute;
 
@@ -44,7 +43,7 @@ public class GrammarPsiImplUtil {
         private List<NavigatablePsiElement> getTargetMethods(String methodName) {
           BnfRule rule = PsiTreeUtil.getParentOfType(getElement(), BnfRule.class);
           String mixinClass = rule == null ? null : getAttribute(rule, KnownAttribute.MIXIN);
-          List<NavigatablePsiElement> implMethods = findRuleImplMethods(javaHelper, psiImplUtilClass, methodName, rule);
+          List<NavigatablePsiElement> implMethods = javaHelper.findRuleImplMethods(psiImplUtilClass, methodName, rule);
           if (!implMethods.isEmpty()) return implMethods;
           List<NavigatablePsiElement> mixinMethods = javaHelper.findClassMethods(mixinClass, JavaHelper.MethodType.INSTANCE, methodName, -1);
           return ContainerUtil.concat(implMethods, mixinMethods);
