+++ docs/architecture.md
@@ -21,9 +21,9 @@ Notable shape: `:bnf-language` and `:jflex-language` are **peers** — neither d
 |-------------------|--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|
 | `:base`           | Shared infrastructure: `GrammarKitBundle` (i18n), `BnfIcons`, message resources.                                                                                                                                                                                                                 |
 | `:parser-runtime` | Language-agnostic parser runtime: `GeneratedParserUtilBase` plus `config.Options`/`config.Option`. No BNF or JFlex specifics.                                                                                                                                                                    |
-| `:bnf-language`   | BNF language model: `psi/`, `parser/`, `analysis/`, `java/`, language registration (`BnfLanguage`, `BnfFileType`, …), `KnownAttribute`, `LightPsi`. Plus PSI-shared rendering and naming primitives (`Renderer`, `JavaRenderer`, `KotlinRenderer`, `Case`, `NameFormat`, `NameShortener`, `JavaNames`, `RuleGraphHelper`, `ParserGeneratorUtil`, `*BnfConstants`) that `:generator` and PSI both consume. Owns `grammars/Grammar.bnf` and BNF `gen/`. |
+| `:bnf-language`   | BNF language model: `psi/`, `parser/`, `java/`, language registration (`BnfLanguage`, `BnfFileType`, …), `KnownAttribute`, `LightPsi`. Plus PSI-shared rendering and naming primitives (`Renderer`, `JavaRenderer`, `KotlinRenderer`, `Case`, `NameFormat`, `NameShortener`, `JavaNames`, `*BnfConstants`) that `:generator` and PSI both consume. Owns `grammars/Grammar.bnf` and BNF `gen/`. |
 | `:jflex-language` | JFlex language support: PSI, parser, editor, file type. True peer of `:bnf-language` — depends only on `:base` and `:parser-runtime`. Owns `grammars/JFlex.bnf` and JFlex `gen/`.                                                                                                                  |
-| `:generator`      | Pure code emission: `Generator`, `JavaParserGenerator`, `KotlinParserGenerator`, `ExpressionGeneratorHelper`, `ExpressionHelper`, `ExpressionInfo`, `RuleMethodsHelper`, `RuleInfo`, `OperatorInfo`, `OperatorType`, `NodeCalls`, `Names`, `FilePrinter`, `OutputOpener`, `GenOptions`, `java/JavaNameShortener`, `kotlin/{KotlinNameShortener, KotlinPlatformConstants}`, plus `Main.java` (headless CLI). Depends on `:bnf-language`. |
+| `:generator`      | Pure code emission: `Generator`, `JavaParserGenerator`, `KotlinParserGenerator`, `ExpressionGeneratorHelper`, `ExpressionHelper`, `ExpressionInfo`, `RuleMethodsHelper`, `RuleInfo`, `OperatorInfo`, `OperatorType`, `NodeCalls`, `Names`, `FilePrinter`, `OutputOpener`, `GenOptions`, `RuleGraphHelper`, `ParserGeneratorUtil`, `analysis/BnfFirstNextAnalyzer`, `java/JavaNameShortener`, `kotlin/{KotlinNameShortener, KotlinPlatformConstants}`, plus `Main.java` (headless CLI). Depends on `:bnf-language`. |
 | Root project      | The IDE plugin itself: `actions/`, `editor/`, `inspection/`, `intention/`, `refactor/`, `search/`, `livePreview/`, `diagram/`, top-level UI providers, `plugin.xml`, signing, publishing.                                                                                                          |
 
 The IntelliJ Platform Gradle plugin (v2) auto-detects subproject dependencies and bundles each module's jar into the plugin distribution.
@@ -54,14 +54,14 @@ The project has a two-tier bootstrap:
 In `:bnf-language`:
 
 - `parser/` — BNF lexer (`GeneratedParserUtilBase` itself lives in `:parser-runtime`, shipped as a template into generated parsers).
-- `psi/` — PSI interfaces/impl plus the extracted BNF-semantic utilities `BnfRules`, `BnfAttributes`, `BnfAst`.
-- `analysis/` — `BnfFirstNextAnalyzer` computes FIRST/FOLLOW sets used by precedence handling, recovery, and inspections.
-- `java/` — `JavaHelper`: ASM-based Java-class introspection for target type resolution.
-- `generator/` (split package, partial) — `Renderer` family, `Case`, `NameFormat`, `NameShortener`, `RuleGraphHelper`, `ParserGeneratorUtil`, `*BnfConstants`, plus `generator/java/{JavaBnfConstants, JavaNames, JavaRenderer}` and `generator/kotlin/{KotlinBnfConstants, KotlinRenderer}`. These are PSI-shared naming/semantic primitives that happen to live in the generator package historically.
+- `psi/` — PSI interfaces/impl plus the BNF-semantic utilities `BnfRules` (rule modifiers, hierarchy, class names), `BnfAttributes` (attribute readers), `BnfAst` (PSI traversal, token maps, `isTokenSequence`, `getChildExpressions`), and `PinMatcher`.
+- `java/` — `JavaHelper`: ASM-based Java-class introspection for target type resolution. Owns `findRuleImplMethods`.
+- `generator/` (split package, partial) — `Renderer` family (with `forPsiClass`/`forPsiImplClass` factories on `NameFormat`), `Case`, `NameFormat`, `NameShortener`, `*BnfConstants`, plus `generator/java/{JavaBnfConstants, JavaNames, JavaRenderer}` and `generator/kotlin/{KotlinBnfConstants, KotlinRenderer}`. These are PSI-shared naming primitives that happen to live in the generator package historically.
 
 In `:generator`:
 
-- `generator/` — Code emission (`Generator`, `Java/KotlinParserGenerator`, `ExpressionGeneratorHelper`, `RuleMethodsHelper`, etc.), output sinks (`FilePrinter`, `OutputOpener`), generator config (`GenOptions`, `Names`), plus `generator/kotlin/{KotlinNameShortener, KotlinPlatformConstants}`.
+- `generator/` — Code emission (`Generator`, `Java/KotlinParserGenerator`, `ExpressionGeneratorHelper`, `RuleMethodsHelper`, etc.), the rule-graph builder `RuleGraphHelper` and emission utilities `ParserGeneratorUtil`, output sinks (`FilePrinter`, `OutputOpener`), generator config (`GenOptions`, `Names`), plus `generator/kotlin/{KotlinNameShortener, KotlinPlatformConstants}`.
+- `analysis/` — `BnfFirstNextAnalyzer` computes FIRST/FOLLOW sets used by precedence handling, recovery, and inspections.
 - `Main.java` — Headless CLI entry point.
 
 In root project (IDE plugin):
@@ -93,4 +93,4 @@ The interesting design choice is the **bootstrap split**: `:bnf-language/gen/` i
 
 - **Aggregate library jar**: `buildGrammarKitJar` currently bundles only root-project classes. To publish a complete `grammar-kit` library to Maven Central post-split, it should also include classes from `:base`, `:parser-runtime`, `:bnf-language`, `:jflex-language`, and `:generator` (e.g., via `zipTree(project(...).tasks.jar.archiveFile)`).
 - **Rename root → `:plugin`**: the root project still serves as the IDE-features module. Splitting it into a dedicated `:plugin` subproject (with a thin coordinator at root) would make the layout fully symmetric, but requires migrating signing/publishing/changelog config.
-- **Tighten the `:bnf-language`/`:generator` split**: today some PSI-shared naming primitives (`Renderer`, `JavaRenderer`, `KotlinRenderer`, `RuleGraphHelper`, `ParserGeneratorUtil`) physically live in `:bnf-language` under the `org.intellij.grammar.generator` package — a deliberate split package. If desired, these could be renamed to `org.intellij.grammar.naming` (or similar) to make module ownership match package boundaries.
+- **Tighten the `:bnf-language`/`:generator` split**: today some PSI-shared naming primitives (`Renderer`, `JavaRenderer`, `KotlinRenderer`) physically live in `:bnf-language` under the `org.intellij.grammar.generator` package — a deliberate split package. If desired, these could be renamed to `org.intellij.grammar.naming` (or similar) to make module ownership match package boundaries.
