+++ generator/src/org/intellij/grammar/generator/ExpressionGeneratorHelper.java
@@ -19,9 +19,9 @@ import org.jetbrains.annotations.Nullable;
 import java.util.*;
 
 import static java.lang.String.format;
-import static org.intellij.grammar.psi.BnfAttributes.*;
-import static org.intellij.grammar.psi.BnfAst.*;
-import static org.intellij.grammar.generator.ParserGeneratorUtil.*;
+import static org.intellij.grammar.generator.ParserGeneratorUtil.ConsumeType;
+import static org.intellij.grammar.generator.ParserGeneratorUtil.quote;
+import static org.intellij.grammar.psi.BnfAttributes.getAttribute;
 
 /**
  * @author greg
