+++ generator/src/org/intellij/grammar/generator/ExpressionHelper.java
@@ -27,10 +27,10 @@ import org.jetbrains.annotations.Nullable;
 import java.util.*;
 import java.util.function.Consumer;
 
-import static org.intellij.grammar.psi.BnfAttributes.*;
-import static org.intellij.grammar.psi.BnfAst.*;
-import static org.intellij.grammar.generator.ParserGeneratorUtil.*;
+import static org.intellij.grammar.generator.ParserGeneratorUtil.topoSort;
 import static org.intellij.grammar.generator.RuleGraphHelper.Cardinality.OPTIONAL;
+import static org.intellij.grammar.psi.BnfAst.getChildExpressions;
+import static org.intellij.grammar.psi.BnfRules.getSuperRules;
 
 /**
  * @author gregsh
